Finished moving the rest of the packets
